% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_depl_fn.R
\name{changeDepl}
\alias{changeDepl}
\title{Changes the depletion value for the depletion year in the data file}
\usage{
changeDepl(dat, para)
}
\arguments{
\item{para}{depletion parameter value for the nth simulation}
}
\description{
Changes the depletion value for the depletion year in the data file
}
\author{
Chantel Wetzel
}
