% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_h_fn.R
\name{changeH}
\alias{changeH}
\title{Changes the steepness value in the control file}
\usage{
changeH(ctl, para)
}
\arguments{
\item{para}{steepness parameter value for the nth simulation}
}
\description{
Changes the steepness value in the control file
}
\author{
Chantel Wetzel
}
