% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_m_fn.R
\name{changeM}
\alias{changeM}
\title{Changes the natural mortality value for female and male fish in the control file}
\usage{
changeM(ctl, para)
}
\arguments{
\item{para}{M vector for female and male values for the nth simulation}
}
\description{
Changes the natural mortality value for female and male fish in the control file
}
\author{
Chantel Wetzel
}
