% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_matrix_fn.R
\name{define_matrix}
\alias{define_matrix}
\title{Create the quant matrix}
\usage{
define_matrix(N, ofl.yrs, depl.yr, n.survey, n.extra.se)
}
\description{
Create the quant matrix
}
\author{
Chantel Wetzel
}
