% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mvt_fn.R
\name{fit.mvt}
\alias{fit.mvt}
\title{This function calculates the covariance between the parameters and resamples based on a multivarite student's t-distribution}
\usage{
fit.mvt(samp.size, para, degree, m.in, depl.in)
}
\description{
This function calculates the covariance between the parameters and resamples based on a multivarite student's t-distribution
}
\author{
Chantel Wetzel
}
