% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_posterior_fn.R
\name{get.new.posterior}
\alias{get.new.posterior}
\title{Calculate the updated posterior probability for the good parameter vectors}
\usage{
get.new.posterior(new.para, degree, m.in, h.in, depl.in)
}
\arguments{
\item{m.in}{initial M values}

\item{h.in}{initial steepness values}

\item{depl.in}{initial depletion values}
}
\description{
Calculate the updated posterior probability for the good parameter vectors
}
\author{
Chantel Wetzel
}
