% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_new_wghts_fn.R
\name{get.new.wghts}
\alias{get.new.wghts}
\title{Calculate the new sampling weights based on the likelihood, prior, and posterior values}
\usage{
get.new.wghts(like, prior, posterior)
}
\description{
Calculate the new sampling weights based on the likelihood, prior, and posterior values
}
\author{
Chantel Wetzel
}
