% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prior_fn.R
\name{get.prior}
\alias{get.prior}
\title{Calculate the prior probabilities for the original good parameter vectors}
\usage{
get.prior(new.para, m.in, h.in, depl.in)
}
\arguments{
\item{m.in}{initial M values}

\item{h.in}{initial steepness values}

\item{depl.in}{initial depletion values}
}
\description{
Calculate the prior probabilities for the original good parameter vectors
}
\author{
Chantel Wetzel
}
