% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seed_fn.R
\name{get.seed}
\alias{get.seed}
\title{Seed Function - this function will generate a seed if one is not read from a file}
\usage{
get.seed()
}
\description{
Seed Function - this function will generate a seed if one is not read from a file
}
\author{
Chantel Wetzel
}
