% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_quant_fn.R
\name{getQuant}
\alias{getQuant}
\title{Function to pull quantities}
\usage{
getQuant(rep.new, parm, ofl.yrs, depl.yr, n.extra.se, n.survey, ssver)
}
\arguments{
\item{ssver}{model version}
}
\description{
Function to pull quantities
}
\author{
Chantel Wetzel
}
