% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_file_fn.R
\name{move.files.fxn}
\alias{move.files.fxn}
\title{Function that moves and saves the needed files
sim.num simulation number that is appended to the report file}
\usage{
move.files.fxn(rep.folder, sim.num)
}
\description{
Function that moves and saves the needed files
sim.num simulation number that is appended to the report file
}
\author{
Chantel Wetzel
}
