% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pars_truncbeta_fn.R
\name{pars.truncbeta}
\alias{pars.truncbeta}
\title{this function gets parameters of standard beta parameterization for depl and Bmsy/B0 from the truncated beta parameters
needed to evaluate densities in AIS}
\usage{
pars.truncbeta(m, s, low, up)
}
\arguments{
\item{m}{}
}
\description{
this function gets parameters of standard beta parameterization for depl and Bmsy/B0 from the truncated beta parameters
needed to evaluate densities in AIS
}
\author{
Chantel Wetzel
}
