% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbeta_ab_fn.R
\name{rbeta.ab}
\alias{rbeta.ab}
\title{Beta Distribution for steepness}
\usage{
rbeta.ab(n, m, s, a, b)
}
\arguments{
\item{n}{= niter}

\item{m}{= depl}

\item{s}{= depl.stdev}

\item{a}{= LB depl}

\item{b}{= UB depl}
}
\description{
Beta Distribution for steepness
}
\author{
Chantel Wetzel
}
