% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fn.R
\name{RepSumFxn}
\alias{RepSumFxn}
\title{Get Final Run Model Output}
\usage{
RepSumFxn(rep.new, n, all.yrs, hist.yrs, ofl.yrs, ssver)
}
\arguments{
\item{rep.new}{report file name to pull summary values from}

\item{n}{simulation number}

\item{all.yrs}{}

\item{hist.yrs}{}

\item{ofl.yrs}{}

\item{ssvers}{}
}
\description{
Get Final Run Model Output
}
\author{
Chantel Wetzel
}
