% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Do_XSSS_AIS.R
\name{SSS.ais.fxn}
\alias{SSS.ais.fxn}
\title{This the main AIS code
The code will run the initial N sample of models
AIS is then started based on initial model weights
The final sample is performed using the weights that exceeded entropy threshold}
\usage{
SSS.ais.fxn(filepath, control.name, dat.name, tantalus = FALSE,
  read.seed = FALSE, entropy.level = 0.92, Niter = 2000,
  AIS.iter = 2000, final.Niter = 5000, m.in, h.in, depl.in)
}
\arguments{
\item{filepath}{= parent directory above the folder with the model files}

\item{control.name}{= paste(file.name,".ctl",sep="")}

\item{dat.name}{= paste(file.name,".dat",sep="")}

\item{tantalus}{= TRUE/FALSE}

\item{read.seed}{= TRUE/FALSE}

\item{Niter}{number of initial model runs}

\item{AIS.iter}{number of AIS model runs}

\item{final.Niter}{number of final model runs, recommened to be greater than N.iter and AIS.iter}

\item{m.in}{= c(prior mean f, prior mean m, prior stdev, prior stdev, start m equal) lognormally distributed}

\item{h.in}{= c(prior mean, prior stdev, lower bound, upper bound) which is a truncated beta}

\item{depl.in}{= c(0.50, 0.20, 0.01, 0.99, distribution)}

\item{hist.yrs}{= the model years; seq(1916,2012,1)}

\item{ofl.yrs}{= years to calculate ofl values; seq(2013,2016,1)}

\item{depl.yr}{Model year associatted with the final index value, does not need to be final model year}
}
\description{
This the main AIS code
The code will run the initial N sample of models
AIS is then started based on initial model weights
The final sample is performed using the weights that exceeded entropy threshold
}
\seealso{
\code{\link{rbeta_ab_fn}}, \code{\link{pars_truncbeta_fn}},
\code{\link{do_sir_fn}}, \code{\link{get_prior_fn}},
\code{\link{get_new_posteriors_fn}}, \code{\link{get_new_wghts_fn}},
\code{\link{fit_mvt_fn}}, \code{\link{change_m_fn}},
\code{\link{change_h_fn}}, \code{\link{change_depl_fn}},
\code{\link{get_quant_fn}}, \code{\link{summary_fn}},
\code{\link{move_file_fn}}, \code{\link{get_see_fn}},
\code{\link{match_fn}}, \code{\link{define_matrix_fn}}
\cod{\linkd{plot_fn}}
}
\author{
Chantel Wetzel
}
